завершив розсортировку
