+++ Web/routes/adminRouter.js
@@ -44,6 +44,9 @@ exports.editgroup_create_post = function (request, response) {
                 })
         );
     }
+    else {
+        response.redirect("/index");
+    }
 };
 
 //Обновляє групу
@@ -57,6 +60,9 @@ exports.updategroup_create_post = function (request, response) {
                     response.redirect("/groups");
                 })
         );
+    }
+    else {
+        response.redirect("/index");
     }
 };
 
@@ -76,6 +82,9 @@ exports.delleteRefId_create_delete = function (request, response) {
                 })
         );
     }
+    else {
+        response.redirect("/index");
+    }
 };
 
 //Удаляє групу
@@ -90,4 +99,61 @@ exports.dellgroup_create_post = function (request, response) {
                 })
         );
     }
+    else {
+        response.redirect("/index");
+    }
 };
+
+//Добавляє групу
+exports.addgroup_create_post = function (request, response) {
+    if (admin.isAdmin(request)) {
+        var data = request.body;
+        var sel = "CALL AddGr('" + data.name + "','" + data.des + "','" + data.date + "','" + data.room + "'," + data.numb + ")";
+        Database.execute(connectionInfo,
+            database => database.query(sel)
+                .then(result => {
+                    response.redirect(request.get('referer'));
+                })
+        );
+    }
+    else {
+        response.redirect("/index");
+    }
+};
+
+//Добавляє урок
+exports.addlesson_create_post = function (request, response) {
+    if (admin.isAdmin(request)) {
+        var data = request.body;
+        var sel = "INSERT INTO lesson( gr_id, dt, room) VALUES ( '" + data.gr_id + "', '" + data.date + "', '" + data.room + "')";
+        Database.execute(connectionInfo,
+            database => database.query(sel)
+                .then(rows => {
+                    response.redirect(request.get('referer'));
+                })
+        );
+    }
+    else {
+        response.redirect("/index");
+    }
+};
+
+//Добавляє дитину
+exports.addchild_create_post = function (request, response) {
+    if (admin.isAdmin(request)) {
+        var data = request.body;
+        var sel = "INSERT INTO child( gr_id, full_name, ref_id) VALUES ( '" + data.group + "', '" + data.name + "', '" + data.id + "')";
+        var sel1 = "DELETE FROM temp WHERE t = '" + data.id + "';";
+        Database.execute(connectionInfo,
+            database => database.query(sel)
+                .then(rows => {
+                    return database.query(sel1);
+                }).then(rows => {
+                    response.redirect(request.get('referer'));
+                })
+        );
+    }
+    else {
+        response.redirect("/index");
+    }
+}
