+++ Web/views/infor_Table.hbs
@@ -85,7 +85,7 @@
                 </button>
             </div>
             <div class="modal-body">
-                <form action="/addgr" method="post" class="w-100">
+                <form action="/addgroup" method="post" class="w-100">
                     <input class="w-100 form-control" maxlength="40" name="name" placeholder="Назва групи" type="text">
                     <input class="w-100 form-control" maxlength="80" name="des" placeholder="Опис групи" type="text">
                     <input class="w-100 form-control" maxlength="10" name="room" placeholder="Кімната" type="text">
