+++ Web/views/main.hbs
@@ -28,7 +28,8 @@
                 dateNow: dateNow,
                 dateEnd: dateEnd,
                 group: group,
-                room: room
+                room: room,
+                order_by: " ORDER BY l.dt ASC"
             });
             Post_CreateTable_index(data,false);
         }
